deleted unnecessary classess, more tests for BugzillaEntriesParser
