+++ src/org/impressivecode/depress/its/ITSDataType.java
@@ -23,8 +23,10 @@ import java.util.List;
 import java.util.Set;
 
 /**
+ * 
  * @author Marek Majchrzak, ImpressiveCode
  * @author Maciej Borkowski, Capgemini Poland
+ * 
  */
 public class ITSDataType {
     private String issueId;
